requirements.txt
salesforce-lavis
openai
flask
flask_sqlalchemy
flask_cors
torch